--- src/mame/machine/spisprit.h
@@ -1,4 +1,0 @@
-// license:BSD-3-Clause
-// copyright-holders:Ville Linde, Nicola Salmoria
-
-void seibuspi_sprite_decrypt(UINT8 *src, int romsize);
