+++ src/mame/video/seibu_crtc.h
@@ -16,6 +16,9 @@
 //  INTERFACE CONFIGURATION MACROS
 //**************************************************************************
 
+#define MCFG_SEIBU_CRTC_DECRYPT_KEY_CB(_devcb) \
+	devcb = &seibu_crtc_device::set_decrypt_key_callback(*device, DEVCB_##_devcb);
+
 #define MCFG_SEIBU_CRTC_LAYER_EN_CB(_devcb) \
 	devcb = &seibu_crtc_device::set_layer_en_callback(*device, DEVCB_##_devcb);
 
@@ -43,6 +46,7 @@ public:
 	// construction/destruction
 	seibu_crtc_device(const machine_config &mconfig, const char *tag, device_t *owner, UINT32 clock);
 
+	template<class _Object> static devcb_base &set_decrypt_key_callback(device_t &device, _Object object) { return downcast<seibu_crtc_device &>(device).m_decrypt_key_cb.set_callback(object); }
 	template<class _Object> static devcb_base &set_layer_en_callback(device_t &device, _Object object) { return downcast<seibu_crtc_device &>(device).m_layer_en_cb.set_callback(object); }
 	template<class _Object> static devcb_base &set_layer_scroll_callback(device_t &device, _Object object) { return downcast<seibu_crtc_device &>(device).m_layer_scroll_cb.set_callback(object); }
 	template<class _Object> static devcb_base &set_reg_1a_callback(device_t &device, _Object object) { return downcast<seibu_crtc_device &>(device).m_reg_1a_cb.set_callback(object); }
@@ -55,7 +59,9 @@ public:
 	DECLARE_READ16_MEMBER( read );
 	DECLARE_READ16_MEMBER( read_alt );
 	DECLARE_READ16_MEMBER( read_xor );
+	DECLARE_WRITE16_MEMBER(decrypt_key_w);
 	DECLARE_WRITE16_MEMBER(layer_en_w);
+	DECLARE_READ16_MEMBER(reg_1a_r);
 	DECLARE_WRITE16_MEMBER(reg_1a_w);
 	DECLARE_WRITE16_MEMBER(layer_scroll_w);
 	DECLARE_WRITE16_MEMBER(layer_scroll_base_w);
@@ -68,6 +74,7 @@ protected:
 	virtual const address_space_config *memory_space_config(address_spacenum spacenum = AS_0) const override;
 
 private:
+	devcb_write16       m_decrypt_key_cb;
 	devcb_write16       m_layer_en_cb;
 	devcb_write16       m_layer_scroll_cb;
 	devcb_write16       m_reg_1a_cb;
@@ -75,6 +82,8 @@ private:
 	const address_space_config      m_space_config;
 	inline UINT16 read_word(offs_t address);
 	inline void write_word(offs_t address, UINT16 data);
+
+	UINT16 m_reg_1a;
 };
